+++ src/pages/index.js
@@ -1,7 +1,7 @@
 import React, { Fragment } from 'react'
 import Helmet from 'react-helmet'
 import Header from 'components/home/Header'
-import About from 'components/About'
+import About from 'components/home/About'
 import Action from 'components/home/Action'
 import Description from 'components/home/Description'
 import Footer from 'components/Footer'
